% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diary.R
\name{dear_diary}
\alias{dear_diary}
\title{Make a diary entry}
\usage{
dear_diary(entry, path = NULL)
}
\arguments{
\item{entry}{The entry you want to make in your diary}

\item{path}{File path corresponding to diary. Default is working directory}
}
\value{
Checks to see whether packages are already installed or loaded. If not installed, will install. If already loaded, will skip loading. Otherwise, will load libraries and provide a quick summary.
}
\description{
Make a diary entry
}
\examples{
\dontrun{
dear_diary("why won't my R code run? :(")
}
}
