% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diary.R
\name{destroy_diary}
\alias{destroy_diary}
\title{Destroy diary}
\usage{
destroy_diary(path = NULL)
}
\arguments{
\item{path}{File path corresponding to diary. Default is working directory}
}
\value{
Delete diary file
}
\description{
Destroy diary
}
\examples{
\dontrun{
destroy_diary()
}

}
