% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libri.R
\name{.check_package_installed}
\alias{.check_package_installed}
\title{Helper: Check to see if packages are installed}
\usage{
.check_package_installed(x)
}
\value{
Logical. TRUE if the package exists, FALSE it not, indicating need for installation.
}
\description{
Helper: Check to see if packages are installed
}
