% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_test.R
\name{lrt_test}
\alias{lrt_test}
\title{Simply likelihood ratio test for comparing two models}
\usage{
lrt_test(m1, m2, df = 1)
}
\arguments{
\item{m1}{a model of class "\code{glm}", "\code{lm}", "\code{nls}", or "\code{Arima}"}

\item{m2}{a model of class "\code{glm}", "\code{lm}", "\code{nls}", or "\code{Arima}". \emph{Assumed to be nested in \code{m1}.}}

\item{df}{degrees of freedom between the two models for testing the test statistic. Default is 1.}
}
\value{
Returns the log-likelihoods from the input models, along with the Chi-sq statistic and p-value from the likelihood ratio test
}
\description{
Simply likelihood ratio test for comparing two models
}
\examples{
\dontrun{
model1 <- lm(mpg ~ wt + hp, data = mtcars)
model2 <- lm(mpg ~ wt, data = mtcars)

lrt_test(m1 = model1, m2 = model2)

}
}
