% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diary.R
\name{new_diary}
\alias{new_diary}
\title{Initialize a diary for recording periodic project notes}
\usage{
new_diary(path = NULL)
}
\arguments{
\item{path}{The file path at which to make a diary. Default is current working directory}
}
\value{
Checks to see whether a diary exists in the working directory and if one does not exist, create one
}
\description{
Initialize a diary for recording periodic project notes
}
\examples{
\dontrun{
new_diary()
}
}
