% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diary.R
\name{read_diary}
\alias{read_diary}
\title{Read diary}
\usage{
read_diary(path = NULL, id = NULL, date = NULL)
}
\arguments{
\item{path}{File path corresponding to diary. Default is working directory}

\item{id}{Read a specific diary entry by id}

\item{date}{Read diary entries on a specific date}
}
\value{
Read diary entries
}
\description{
Read diary
}
\examples{
\dontrun{
read_diary()
}

}
