% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_function.R
\name{use_r_function}
\alias{use_r_function}
\title{Save a function in the global environment to an R file}
\usage{
use_r_function(function_object, path = NULL)
}
\arguments{
\item{function_object}{The name of the function object in the Global environment}

\item{path}{Directory to save function script in. Default is working directory}
}
\value{
Checks to see whether packages are already installed or loaded. If not installed, will install. If already loaded, will skip loading. Otherwise, will load libraries and provide a quick summary.
}
\description{
Save a function in the global environment to an R file
}
\examples{
\dontrun{
hello <- function() { print("hello, world!")}

use_r_function("hello")

}
}
