% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ally-package.R
\docType{package}
\name{ally-package}
\alias{ally}
\alias{ally-package}
\title{ally: Helpers for using AI coding tools with R}
\description{
Tooling that helps R users get the most out of AI coding assistants.
The first feature set is skill installation: download a skill from
GitHub (or copy from a local path) and link it into every detected
agent's expected folder.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rfortherestofus/ally}
  \item Report bugs at \url{https://github.com/rfortherestofus/ally/issues}
}

}
\author{
\strong{Maintainer}: David Keyes \email{david@rfortherestofus.com}

Authors:
\itemize{
  \item David Keyes \email{david@rfortherestofus.com}
}

}
\keyword{internal}
